Refactor out UnitCard. Apply styles for desktop
